+++ src/test/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/data/RedshiftDataApiRepositoryTest.kt
@@ -2,11 +2,11 @@ package uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.params.ParameterizedTest
+import org.junit.jupiter.params.provider.CsvSource
 import org.mockito.Mockito.mock
-import org.mockito.Mockito.verify
 import org.mockito.kotlin.any
 import org.mockito.kotlin.eq
-import org.mockito.kotlin.times
 import org.mockito.kotlin.whenever
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate
@@ -15,13 +15,12 @@ import software.amazon.awssdk.services.redshiftdata.model.DescribeStatementReque
 import software.amazon.awssdk.services.redshiftdata.model.DescribeStatementResponse
 import software.amazon.awssdk.services.redshiftdata.model.ExecuteStatementRequest
 import software.amazon.awssdk.services.redshiftdata.model.ExecuteStatementResponse
-import software.amazon.awssdk.services.redshiftdata.model.SqlParameter
-import software.amazon.awssdk.services.redshiftdata.model.ValidationException
-import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.ConfiguredApiController.FiltersPrefix.RANGE_FILTER_END_SUFFIX
-import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.ConfiguredApiController.FiltersPrefix.RANGE_FILTER_START_SUFFIX
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.ConfiguredApiRepositoryTest.Companion.REPOSITORY_TEST_POLICY_ENGINE_RESULT
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.ConfiguredApiRepositoryTest.Companion.REPOSITORY_TEST_QUERY
+import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.RepositoryHelper.Companion.FALSE_WHERE_CLAUSE
+import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.RepositoryHelper.Companion.TRUE_WHERE_CLAUSE
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.RepositoryHelper.FilterType
+import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.policyengine.Policy.PolicyResult
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.redshiftdata.StatementExecutionResponse
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.redshiftdata.StatementExecutionStatus
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service.TableIdGenerator
@@ -39,7 +38,7 @@ class RedshiftDataApiRepositoryTest {
                   WITH dataset_ AS (SELECT prisoners.number AS prisonNumber,CONCAT(CONCAT(prisoners.lastname, ', '), substring(prisoners.firstname, 1, 1)) AS name,movements.time AS date,movements.direction,movements.type,movements.origin,movements.origin_code,movements.destination,movements.destination_code,movements.reason
         FROM datamart.domain.movement_movement as movements
         JOIN datamart.domain.prisoner_prisoner as prisoners
-        ON movements.prisoner = prisoners.id),policy_ AS (SELECT * FROM dataset_ WHERE (origin_code IN ('HEI','LWSTMC','NSI','LCI','TCI') AND lower(direction)='out') OR (destination_code IN ('HEI','LWSTMC','NSI','LCI','TCI') AND lower(direction)='in')),filter_ AS (SELECT * FROM policy_ WHERE lower(direction) = :direction)
+        ON movements.prisoner = prisoners.id),policy_ AS (SELECT * FROM dataset_ WHERE (origin_code IN ('HEI','LWSTMC','NSI','LCI','TCI') AND lower(direction)='out') OR (destination_code IN ('HEI','LWSTMC','NSI','LCI','TCI') AND lower(direction)='in')),filter_ AS (SELECT * FROM policy_ WHERE lower(direction) = 'out')
         SELECT *
                   FROM filter_ ORDER BY date asc
                   );
@@ -47,27 +46,31 @@ class RedshiftDataApiRepositoryTest {
 
     val movementPrisoner1 = mapOf("id" to "171034.12", "prisoner" to 171034L, "date" to LocalDateTime.of(2010, 12, 17, 0, 0, 0), "time" to LocalDateTime.of(2010, 12, 17, 7, 12, 0), "direction" to "OUT", "type" to "CRT", "origin_code" to "LFI", "origin" to "LANCASTER FARMS (HMPYOI)", "destination_code" to "STHEMC", "destination" to "St. Helens Magistrates Court", "reason" to "Production (Sentence/Civil Custody)")
     val movementPrisoner2 = mapOf("id" to "227482.1", "prisoner" to 227482L, "date" to LocalDateTime.of(2010, 12, 8, 0, 0, 0), "time" to LocalDateTime.of(2010, 12, 8, 10, 8, 0), "direction" to "IN", "type" to "ADM", "origin_code" to "IMM", "origin" to "Immigration", "destination_code" to "HRI", "destination" to "Haslar Immigration Removal Centre", "reason" to "Detained Immigration Act 71 -Wait Deport")
+    private const val REDSHIFT_DATA_API_DB: String = "redshiftDB"
+    private const val REDSHIFT_DATA_API_CLUSTER_ID: String = "redshiftClusterId"
+    private const val REDSHIFT_DATA_API_SECRET_ARN: String = "redshiftSecretArn"
   }
 
   @Test
   fun `executeQueryAsync should call the redshift data api with the correct query and return the execution id and table id`() {
     val redshiftDataClient = mock<RedshiftDataClient>()
-    val executeStatementRequestBuilder = mock<ExecuteStatementRequest.Builder>()
     val executeStatementResponse = mock<ExecuteStatementResponse>()
     val tableIdGenerator = mock<TableIdGenerator>()
     val tableId = "_a6227417_bdac_40bb_bc81_49c750daacd7"
     val executionId = "someId"
     val redshiftDataApiRepository = RedshiftDataApiRepository(
       redshiftDataClient,
-      executeStatementRequestBuilder,
       tableIdGenerator,
+      REDSHIFT_DATA_API_DB,
+      REDSHIFT_DATA_API_CLUSTER_ID,
+      REDSHIFT_DATA_API_SECRET_ARN,
     )
-
-    val mockedBuilderWithSql = ExecuteStatementRequest.builder()
-      .clusterIdentifier("ab")
-      .database("cd")
-      .secretArn("ef")
+    val executeStatementRequest = ExecuteStatementRequest.builder()
+      .clusterIdentifier(REDSHIFT_DATA_API_CLUSTER_ID)
+      .database(REDSHIFT_DATA_API_DB)
+      .secretArn(REDSHIFT_DATA_API_SECRET_ARN)
       .sql(sqlStatement(tableId))
+      .build()
 
     whenever(
       tableIdGenerator.generateNewExternalTableId(),
@@ -76,24 +79,8 @@ class RedshiftDataApiRepositoryTest {
     )
 
     whenever(
-      executeStatementRequestBuilder.sql(
-        sqlStatement(tableId),
-      ),
-    ).thenReturn(
-      mockedBuilderWithSql,
-    )
-
-    val queryParams = listOf(SqlParameter.builder().name("direction").value("out").build())
-    whenever(
-      executeStatementRequestBuilder.parameters(queryParams),
-    ).thenReturn(
-      mockedBuilderWithSql
-        .parameters(queryParams),
-    )
-
-    whenever(
       redshiftDataClient.executeStatement(
-        mockedBuilderWithSql.build(),
+        executeStatementRequest,
       ),
     ).thenReturn(executeStatementResponse)
 
@@ -113,13 +100,13 @@ class RedshiftDataApiRepositoryTest {
   }
 
   @Test
-  fun `executeQueryAsync should replace filters which contain a dot in their name with an underscore and not error`() {
-    val dateStartKeyUnderscore = "date_start"
-    val dateEndKeyUnderscore = "date_end"
+  fun `executeQueryAsync should call the redshift data api with the correct query including all filters and return the execution id and table id`() {
     val startDate = "2024-02-16"
     val endDate = "2024-02-17"
+    val directionFilter = ConfiguredApiRepository.Filter("direction", "out")
     val startDateFilter = ConfiguredApiRepository.Filter("date", startDate, FilterType.DATE_RANGE_START)
     val endDateFilter = ConfiguredApiRepository.Filter("date", endDate, FilterType.DATE_RANGE_END)
+    val nameDynamicFilter = ConfiguredApiRepository.Filter("name", "LastNa", FilterType.DYNAMIC)
     val tableId = "_a6227417_bdac_40bb_bc81_49c750daacd7"
     val executionId = "someId"
     val sqlStatement =
@@ -130,26 +117,27 @@ class RedshiftDataApiRepositoryTest {
           WITH dataset_ AS (SELECT prisoners.number AS prisonNumber,CONCAT(CONCAT(prisoners.lastname, ', '), substring(prisoners.firstname, 1, 1)) AS name,movements.time AS date,movements.direction,movements.type,movements.origin,movements.origin_code,movements.destination,movements.destination_code,movements.reason
 FROM datamart.domain.movement_movement as movements
 JOIN datamart.domain.prisoner_prisoner as prisoners
-ON movements.prisoner = prisoners.id),policy_ AS (SELECT * FROM dataset_ WHERE (origin_code IN ('HEI','LWSTMC','NSI','LCI','TCI') AND lower(direction)='out') OR (destination_code IN ('HEI','LWSTMC','NSI','LCI','TCI') AND lower(direction)='in')),filter_ AS (SELECT * FROM policy_ WHERE date >= CAST(:$dateStartKeyUnderscore AS timestamp) AND date < (CAST(:$dateEndKeyUnderscore AS timestamp) + INTERVAL '1' day))
+ON movements.prisoner = prisoners.id),policy_ AS (SELECT * FROM dataset_ WHERE (origin_code IN ('HEI','LWSTMC','NSI','LCI','TCI') AND lower(direction)='out') OR (destination_code IN ('HEI','LWSTMC','NSI','LCI','TCI') AND lower(direction)='in')),filter_ AS (SELECT * FROM policy_ WHERE date >= CAST('$startDate' AS timestamp) AND date < (CAST('$endDate' AS timestamp) + INTERVAL '1' day) AND lower(direction) = 'out' AND name ILIKE 'LastNa%')
 SELECT *
           FROM filter_ ORDER BY date asc
           );
       """.trimIndent()
     val redshiftDataClient = mock<RedshiftDataClient>()
-    val executeStatementRequestBuilder = mock<ExecuteStatementRequest.Builder>()
     val executeStatementResponse = mock<ExecuteStatementResponse>()
     val tableIdGenerator = mock<TableIdGenerator>()
     val redshiftDataApiRepository = RedshiftDataApiRepository(
       redshiftDataClient,
-      executeStatementRequestBuilder,
       tableIdGenerator,
+      REDSHIFT_DATA_API_DB,
+      REDSHIFT_DATA_API_CLUSTER_ID,
+      REDSHIFT_DATA_API_SECRET_ARN,
     )
-
-    val mockedBuilderWithSql = ExecuteStatementRequest.builder()
-      .clusterIdentifier("ab")
-      .database("cd")
-      .secretArn("ef")
+    val executeStatementRequest = ExecuteStatementRequest.builder()
+      .clusterIdentifier(REDSHIFT_DATA_API_CLUSTER_ID)
+      .database(REDSHIFT_DATA_API_DB)
+      .secretArn(REDSHIFT_DATA_API_SECRET_ARN)
       .sql(sqlStatement)
+      .build()
 
     whenever(
       tableIdGenerator.generateNewExternalTableId(),
@@ -158,35 +146,8 @@ SELECT *
     )
 
     whenever(
-      executeStatementRequestBuilder.sql(
-        sqlStatement,
-      ),
-    ).thenReturn(
-      mockedBuilderWithSql,
-    )
-
-    val queryParamsWithUnderscores = listOf(
-      SqlParameter.builder().name(dateStartKeyUnderscore).value(startDate).build(),
-      SqlParameter.builder().name(dateEndKeyUnderscore).value(endDate).build(),
-    )
-    whenever(
-      executeStatementRequestBuilder.parameters(queryParamsWithUnderscores),
-    ).thenReturn(
-      mockedBuilderWithSql
-        .parameters(queryParamsWithUnderscores),
-    )
-
-    val queryParamsWithDots = listOf(
-      SqlParameter.builder().name("date$RANGE_FILTER_START_SUFFIX").value(startDate).build(),
-      SqlParameter.builder().name("date$RANGE_FILTER_END_SUFFIX").value(endDate).build(),
-    )
-    whenever(
-      executeStatementRequestBuilder.parameters(queryParamsWithDots),
-    ).thenThrow(ValidationException::class.java)
-
-    whenever(
       redshiftDataClient.executeStatement(
-        mockedBuilderWithSql.build(),
+        executeStatementRequest,
       ),
     ).thenReturn(executeStatementResponse)
 
@@ -199,6 +160,8 @@ SELECT *
       filters = listOf(
         startDateFilter,
         endDateFilter,
+        directionFilter,
+        nameDynamicFilter,
       ),
       sortColumn = "date",
       sortedAsc = true,
@@ -214,8 +177,10 @@ SELECT *
     val tableIdGenerator = mock<TableIdGenerator>()
     val redshiftDataApiRepository = RedshiftDataApiRepository(
       redshiftDataClient,
-      mock(),
       tableIdGenerator,
+      REDSHIFT_DATA_API_DB,
+      REDSHIFT_DATA_API_CLUSTER_ID,
+      REDSHIFT_DATA_API_SECRET_ARN,
     )
     val statementId = "statementId"
     val status = "FINISHED"
@@ -254,13 +219,14 @@ SELECT *
   @Test
   fun `executeQueryAsync should call the redshift data api and not error when no filters are provided`() {
     val redshiftDataClient = mock<RedshiftDataClient>()
-    val executeStatementRequestBuilder = mock<ExecuteStatementRequest.Builder>()
     val executeStatementResponse = mock<ExecuteStatementResponse>()
     val tableIdGenerator = mock<TableIdGenerator>()
     val redshiftDataApiRepository = RedshiftDataApiRepository(
       redshiftDataClient,
-      executeStatementRequestBuilder,
       tableIdGenerator,
+      REDSHIFT_DATA_API_DB,
+      REDSHIFT_DATA_API_CLUSTER_ID,
+      REDSHIFT_DATA_API_SECRET_ARN,
     )
     val tableId = "_a6227417_bdac_40bb_bc81_49c750daacd7"
     val executionId = "someId"
@@ -273,17 +239,18 @@ SELECT *
                   WITH dataset_ AS (SELECT prisoners.number AS prisonNumber,CONCAT(CONCAT(prisoners.lastname, ', '), substring(prisoners.firstname, 1, 1)) AS name,movements.time AS date,movements.direction,movements.type,movements.origin,movements.origin_code,movements.destination,movements.destination_code,movements.reason
         FROM datamart.domain.movement_movement as movements
         JOIN datamart.domain.prisoner_prisoner as prisoners
-        ON movements.prisoner = prisoners.id),policy_ AS (SELECT * FROM dataset_ WHERE (origin_code IN ('HEI','LWSTMC','NSI','LCI','TCI') AND lower(direction)='out') OR (destination_code IN ('HEI','LWSTMC','NSI','LCI','TCI') AND lower(direction)='in')),filter_ AS (SELECT * FROM policy_ WHERE TRUE)
+        ON movements.prisoner = prisoners.id),policy_ AS (SELECT * FROM dataset_ WHERE (origin_code IN ('HEI','LWSTMC','NSI','LCI','TCI') AND lower(direction)='out') OR (destination_code IN ('HEI','LWSTMC','NSI','LCI','TCI') AND lower(direction)='in')),filter_ AS (SELECT * FROM policy_ WHERE $TRUE_WHERE_CLAUSE)
         SELECT *
                   FROM filter_ ORDER BY date asc
                   );
       """.trimIndent()
 
-    val mockedBuilderWithSql = ExecuteStatementRequest.builder()
-      .clusterIdentifier("ab")
-      .database("cd")
-      .secretArn("ef")
+    val executeStatementRequest = ExecuteStatementRequest.builder()
+      .clusterIdentifier(REDSHIFT_DATA_API_CLUSTER_ID)
+      .database(REDSHIFT_DATA_API_DB)
+      .secretArn(REDSHIFT_DATA_API_SECRET_ARN)
       .sql(finalQuery)
+      .build()
 
     whenever(
       tableIdGenerator.generateNewExternalTableId(),
@@ -292,21 +259,8 @@ SELECT *
     )
 
     whenever(
-      executeStatementRequestBuilder.sql(
-        finalQuery,
-      ),
-    ).thenReturn(
-      mockedBuilderWithSql,
-    )
-
-    val queryParams = emptyList<SqlParameter>()
-    whenever(
-      executeStatementRequestBuilder.parameters(queryParams),
-    ).thenThrow(ValidationException::class.java)
-
-    whenever(
       redshiftDataClient.executeStatement(
-        mockedBuilderWithSql.build(),
+        executeStatementRequest,
       ),
     ).thenReturn(executeStatementResponse)
 
@@ -322,7 +276,72 @@ SELECT *
       policyEngineResult = REPOSITORY_TEST_POLICY_ENGINE_RESULT,
     )
 
-    verify(executeStatementRequestBuilder, times(0)).parameters(any<List<SqlParameter>>())
+    assertEquals(StatementExecutionResponse(tableId, executionId), actual)
+  }
+
+  @ParameterizedTest
+  @CsvSource(
+    "${PolicyResult.POLICY_PERMIT}, $TRUE_WHERE_CLAUSE",
+    "${PolicyResult.POLICY_DENY}, $FALSE_WHERE_CLAUSE",
+  )
+  fun `executeQueryAsync should call the redshift data api with the converted sql where clause for permit and denied policies`(policyEngineResult: String, queryWhereClause: String) {
+    val redshiftDataClient = mock<RedshiftDataClient>()
+    val executeStatementResponse = mock<ExecuteStatementResponse>()
+    val tableIdGenerator = mock<TableIdGenerator>()
+    val redshiftDataApiRepository = RedshiftDataApiRepository(
+      redshiftDataClient,
+      tableIdGenerator,
+      REDSHIFT_DATA_API_DB,
+      REDSHIFT_DATA_API_CLUSTER_ID,
+      REDSHIFT_DATA_API_SECRET_ARN,
+    )
+    val tableId = "_a6227417_bdac_40bb_bc81_49c750daacd7"
+    val executionId = "someId"
+    val finalQuery =
+      """
+                  CREATE EXTERNAL TABLE reports.$tableId 
+                  STORED AS parquet 
+                  LOCATION 's3://dpr-working-development/reports/$tableId/' 
+                  AS ( 
+                  WITH dataset_ AS (SELECT prisoners.number AS prisonNumber,CONCAT(CONCAT(prisoners.lastname, ', '), substring(prisoners.firstname, 1, 1)) AS name,movements.time AS date,movements.direction,movements.type,movements.origin,movements.origin_code,movements.destination,movements.destination_code,movements.reason
+        FROM datamart.domain.movement_movement as movements
+        JOIN datamart.domain.prisoner_prisoner as prisoners
+        ON movements.prisoner = prisoners.id),policy_ AS (SELECT * FROM dataset_ WHERE $queryWhereClause),filter_ AS (SELECT * FROM policy_ WHERE $TRUE_WHERE_CLAUSE)
+        SELECT *
+                  FROM filter_ ORDER BY date asc
+                  );
+      """.trimIndent()
+
+    val executeStatementRequest = ExecuteStatementRequest.builder()
+      .clusterIdentifier(REDSHIFT_DATA_API_CLUSTER_ID)
+      .database(REDSHIFT_DATA_API_DB)
+      .secretArn(REDSHIFT_DATA_API_SECRET_ARN)
+      .sql(finalQuery)
+      .build()
+
+    whenever(
+      tableIdGenerator.generateNewExternalTableId(),
+    ).thenReturn(
+      tableId,
+    )
+
+    whenever(
+      redshiftDataClient.executeStatement(
+        executeStatementRequest,
+      ),
+    ).thenReturn(executeStatementResponse)
+
+    whenever(
+      executeStatementResponse.id(),
+    ).thenReturn(executionId)
+
+    val actual = redshiftDataApiRepository.executeQueryAsync(
+      query = REPOSITORY_TEST_QUERY,
+      filters = emptyList(),
+      sortColumn = "date",
+      sortedAsc = true,
+      policyEngineResult = policyEngineResult,
+    )
 
     assertEquals(StatementExecutionResponse(tableId, executionId), actual)
   }
@@ -335,8 +354,10 @@ SELECT *
     val jdbcTemplate = mock<NamedParameterJdbcTemplate>()
     val redshiftDataApiRepository = RedshiftDataApiRepository(
       redshiftDataClient,
-      mock(),
       tableIdGenerator,
+      REDSHIFT_DATA_API_DB,
+      REDSHIFT_DATA_API_CLUSTER_ID,
+      REDSHIFT_DATA_API_SECRET_ARN,
     )
     val selectedPage = 1L
     val pageSize = 10L
@@ -361,7 +382,9 @@ SELECT *
     val redshiftDataApiRepository = RedshiftDataApiRepository(
       mock(),
       mock(),
-      mock(),
+      REDSHIFT_DATA_API_DB,
+      REDSHIFT_DATA_API_CLUSTER_ID,
+      REDSHIFT_DATA_API_SECRET_ARN,
     )
     val expected = listOf<Map<String, Any?>>(mapOf("total" to 5L))